+++ admin/App.jsx
@@ -3,14 +3,14 @@ import LoginForm from './components/LoginForm';
 import Dashboard from './components/Dashboard';
 
 const App = () => {
-  const [screen, setScreen] = useState('Login');
+  const [screen, setScreen] = useState({val: 'Login', params: {}});
 
   return (
     <>
       { 
-        (screen == 'Dashboard') ? 
-            <Dashboard setScreen={setScreen}/> : 
-            <LoginForm setScreen={setScreen}/>
+        (screen.val == 'Dashboard') ? 
+            <Dashboard params={screen.params} setScreen={setScreen}/> : 
+            <LoginForm params={screen.params} setScreen={setScreen}/>
       }
     </>
   );